requirements.txt
langchain
langchain-community
langchain-openai
openai
faiss-cpu
python-dateutil
pydantic
python-dotenv
dateparser
tqdm
docx2txt
PyPDF2
unstructured
lxml
html2text
bs4
tabulate
typing-extensions
